feat: bind keys only when necessary
